@@ -15,18 +15,22 @@
         }
     },
     "config": {
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": true
+        }
     },
     "require": {
-        "php": "^7.2 || ^8.0",
-        "symfony/event-dispatcher": "^4.4 || ^5.3",
-        "symfony/http-kernel": "^4.4 || ^5.3",
-        "symfony/messenger": "^4.4 || ^5.3"
+        "php": "^7.4 || ^8.0",
+        "symfony/event-dispatcher": "^5.4 || ^6.0",
+        "symfony/http-kernel": "^5.4 || ^6.0",
+        "symfony/messenger": "^5.4 || ^6.0"
     },
     "require-dev": {
         "myonlinestore/coding-standard": "^3.1",
-        "phpunit/phpunit": "^8.5 || ^9.5",
+        "myonlinestore/php-devtools": "^0.2.3",
+        "phpunit/phpunit": "^9.5",
         "psalm/plugin-phpunit": "^0.16",
-        "vimeo/psalm": "^4.9"
+        "vimeo/psalm": "^4.22"
     }
 }
